Refactor test handlers out into dedicated file
